checkpoint: extracts `toShortfinBatchedRequestBody` into module
