+++ src/library/ShimmedStabilityAIClient/conversions/exports.ts
@@ -1,3 +1,5 @@
 export * from './toShortfinRequestBodyPrompt';
 
 export * from './toShortfinRequestBody';
+
+export * from './toShortfinBatchedRequestBody';
