+++ src/library/ShimmedStabilityAIClient/definition.ts
@@ -10,7 +10,6 @@ import type {
 import Base64CharacterEncodedByteSequence from '@/library/Base64CharacterEncodedByteSequence';
 import HTTP from '@/library/HTTP';
 import Schema from '@/library/Schema';
-import Shortfin from '@/library/Shortfin';
 
 import {
   URLOrigin,
@@ -18,24 +17,8 @@ import {
 } from '@/library/URLComponent';
 
 import {
-  toShortfinRequestBody,
+  toShortfinBatchedRequestBody,
 } from './conversions';
-
-const toShortfinBatchedRequestBody = (
-  givenRequestBodies: GenerateFromTextRequest['textToImageRequestBody'][],
-): Shortfin.TextToImage.SDXL.Client.Request.Body.Batched => {
-  const derivedShortfinRequestBodies = givenRequestBodies
-    .map(toShortfinRequestBody)
-    .filter($0 => $0 !== null);
-
-  const derivedBatchedRequestBody = new Shortfin.TextToImage.SDXL.Client.Request.Body.Batched();
-
-  derivedShortfinRequestBodies.forEach((eachShortfinRequestBody) => {
-    derivedBatchedRequestBody.append(eachShortfinRequestBody);
-  });
-
-  return derivedBatchedRequestBody;
-};
 
 const Shortfin_TextToImage_SDXL_Client_Response_Body = {
   SchemaMember: {
